Migrate Organization admin tests into new testsuite
